Improve support for 1D computations
Add mplementation of output_manager for 1D data. As opposed to higher
dimensions, it uses gnuplot output (simple list of numbers), since the
VTK format and Paraview don't seem to be useful for 1D data.

Gnuplot output for 1D data was bugged and raw output used incorrect way
of indexing the vector ([] instead of () as for tensors). Fix both of
these issues.

Evaluation of 1D B-splines was inconsistent with the multidimensional
functions. Unify the evaluation functionality and change name of the
file accordingly.