Fix crashing StatTimer initialization
In more recent versions of Galois, StatTimer requires initialized
ThreadPool, since it uses specialized per-thread allocators. This can be
done by creating galois_executor. As long as galois_executor is created
prior to StatTimer, everything is fine. In some example applications
this order was reversed. That was not a problem when using Galois 2.2.1
and was not noticed during migration to Galois 6.0.

Closes #30