Add .clang-tidy file with reasonable settings
This configuration results in a lot of warnings. Silencing all of these
will take considerable effort, but should ultimately improve the quality
of the code.

One issue that remains open are the naming conventions. Most of the code
follows the lowercase convention, but for some entities it really makes
sense to use original mathematical names (test and trial spaces U and V,
norms - L2, H1).

Closes #7