Ensure git commands are executed inside repository
When cmake was invoked outside the project directory, git rev-parse
would fail since it was executed outside a git repo (or worse - if the
invocation directory happened to be a git repo, it would silently fail
by providing a wrong SHA). This change ensures git is invoked in the
correct directory.

Closes #36