Use feature macros to simplify the build process
Since Galois and MUMPS availability is now visible in the code, the
logic to exclude contents of relevant files can be moved from the build
system (conditionally adding source files) to the preprocessor.

This is also beneficial to the consumers of the library, since using
unavailable features will now result in compiler complaining about
missing names instead of missing includes.