Add an intermediate object library
This makes it easier to get reasonable coverage statistics, since when
test suite is linked with the static library, unused object files are
omitted and are thus missing from the coverage stats. This does not
happen when linking to the object library.

An alternative solution would be to use --whole-archive linker flag:
  -Wl,--whole-archive ADS -Wl,--no-whole-archive
when linking the (non-object) library.