Add copy assignment operator to basis_data
This makes the class consistent from the point of view of memory
management. Move operations are not declared, and thus the moves fall
back to copy operations. This is desirable for now, until manual memory
management of basis_data is replaced with proper multi-dimensional
arrays that mange their own memory.

This does not follow the Core Guideline C.21 and is picked up by
clang-tidy, but it is cleaner than explicitly declaring the move
operations and implementing these in terms of moves.