Avoid calling operator () by (*this)(...)
This leads to issues with clang-format, since 12.0.0 formats it as
(*this) (...), while 12.0.1 and higher like (*this)(...) (no space).  In
the code that uses this construction, operator () is used strictly to
make interface more pleasant, so to make the implementation more
readable it is a good idea anyway to give this operation a meaningful
name.