Simplify high order correction in multistep code
In this method we start with a linear multistep scheme of the general
form

  (M + eta K) u[n] = RHS

We want to replace (M + eta K) with (Mx + eta Kx) @ (My + eta Ky), where
@ denotes the Kronecker product, but this would introduce too high an
error. To prevent that, we modify the equation so that instead of u[n]
we solve for something smaller, e.g.

  (M + eta K) (u[n] - 2 u[n-1] + u[n-2])
  =
  RHS + (M + eta K) (-2 u[n-1] + u[n-2])

After replacing (M + eta K) with the aforementioned approximation, we
can move everything except for u[n] to the right-hand side:

  (M + eta K) u[n] = RHS - eta^2 Kx @ Ky (-2 u[n-1] + u[n-2])

This turns out to be a bad idea though, since the bilinear form
associated with Kx @ Ky is a messy and unnatural expression that
involves mixed derivatives. Instead, it is better to leave it as it is,
solve for U = u[n] - 2 u[n-1] + u[n-2] and compute u[n] from U.