Ensure correct linking with static MUMPS
MUMPS needs LAPACK and BLAS. In case it is included as a shared library,
CMake doesn't need to know about it, since order of libraries on in the
linker invocation is irrelevant. If MUMPS is used as a static library
though, we must make sure BLAS and LAPACK are included after it, so that
the linker can correctly resolve all symbols.

Some other issues related to static libraries are also fixed here:

 - ScaLAPACK was typically not found by CMake, and thus not explicitly
   included on the list of libraries, since on Debian systems it is
   usually present as libscalapack-openmpi.

 - Not all necessary ordering libraries were being sought

Closes #66