Make refinement preserve aspect ratio
In the simple refinement scheme for the higher-order CIP method
refinements of the regular mesh are performed exclusively in the x
direction, leading to extremely elongated elements near the boundary
layer, and consequently, to numerical issues. This change introduces
refinements in the y direction to preserve some fixed upper bound on the
aspect ratio of the elements.